[REFACTOR] Document 생성 Response 수정
단일 생성에서 복수 생성으로 수정
